Move CR printing from print_page_residency_chart.
The carriage return is not needed the first time
print_page_residency_chart() is called.  By moving \r printing to the
caller, it knows which is the first time.

Only within the o_touch loop is print_page_residency_chart() called
multiple times, so only within that loop is \r necessary.

The final call to print_page_residency_chart() has been removed since it
would print identical content to the previous time it was called,
whether than be within the o_touch loop or not.  Hence it has been
removed as all that's needed is to terminate the line with \n.